requirements.txt
flask
python-dotenv
openai-whisper
openai>=1.0.0
flask-limiter
